+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## ChangeLog for: chainjs
 
+## Version 0.1.0 - 2014/11/17
+
+- [APIs]: remove "filter", "before", "sham" methods, and use "destroy" instead of "stop"
+- [Features]: Chain refactor using Class Constructor for less memory usage. The code is more clear.
+
 ## Version 0.0.3-2 - 2014/1/17
 
 - [Fix bug]: data() api using extend to export chain data instead of return private data variable